Update Seven Languages to Xtext 2.11

+++ languages/org.xtext.builddsl.lib/xtend-gen/org/xtext/builddsl/lib/DirectoryCleaner.java
@@ -8,27 +8,23 @@
 package org.xtext.builddsl.lib;
 
 import java.io.File;
+import java.util.List;
+import java.util.function.Consumer;
 import org.eclipse.xtext.xbase.lib.Conversions;
-import org.eclipse.xtext.xbase.lib.IterableExtensions;
-import org.eclipse.xtext.xbase.lib.Procedures.Procedure1;
 
 @SuppressWarnings("all")
 public class DirectoryCleaner {
   public static void deleteDirectoryContents(final File file) {
     boolean _isDirectory = file.isDirectory();
     if (_isDirectory) {
-      File[] _listFiles = file.listFiles();
-      final Procedure1<File> _function = new Procedure1<File>() {
-        @Override
-        public void apply(final File it) {
-          boolean _isDirectory = it.isDirectory();
-          if (_isDirectory) {
-            DirectoryCleaner.deleteDirectoryContents(it);
-          }
-          it.delete();
+      final Consumer<File> _function = (File it) -> {
+        boolean _isDirectory_1 = it.isDirectory();
+        if (_isDirectory_1) {
+          DirectoryCleaner.deleteDirectoryContents(it);
         }
+        it.delete();
       };
-      IterableExtensions.<File>forEach(((Iterable<File>)Conversions.doWrapArray(_listFiles)), _function);
+      ((List<File>)Conversions.doWrapArray(file.listFiles())).forEach(_function);
     } else {
       file.delete();
     }
